test(bpmn-model): add more validation tests for message catch element
